requirements.txt
pypdf>=4.0.0
python-docx>=1.0.0
python-dotenv>=1.0.0
openai>=1.0.0
langsmith>=0.1.0